+++ propagator/opentelemetry-propagator-aws-xray/src/opentelemetry/propagators/aws/__init__.py
@@ -47,7 +47,7 @@ Or by setting this propagator in your instrumented application:
 .. code-block:: python
 
     from opentelemetry.propagate import set_global_textmap
-    from opentelemetry.sdk.extension.aws.trace.propagation.aws_xray_format import AwsXRayFormat
+    from opentelemetry.propagators.aws import AwsXRayFormat
 
     set_global_textmap(AwsXRayFormat())